GC - Fixing ts issues
