+++ package.json
@@ -6,7 +6,8 @@
     "web": "expo start --web",
     "start": "react-native start",
     "test": "jest",
-    "expo": "expo start"
+    "expo": "expo start",
+    "preinstall": "npx npm-force-resolutions"
   },
   "dependencies": {
     "@expo-google-fonts/inter": "^0.1.0",
@@ -25,15 +26,7 @@
     "react": "16.13.1",
     "react-dom": "16.13.1",
     "react-native": "^0.64.0",
-    "react-native-advance-draggable-view": "^1.0.13",
-    "react-native-deck-swiper": "^2.0.5",
-    "react-native-draggable": "^3.3.0",
     "react-native-elements": "^3.3.1",
-    "react-native-gesture-handler": "~1.8.0",
-    "react-native-paper": "^4.7.2",
-    "react-native-reanimated": "~1.13.0",
-    "react-native-safe-area-context": "3.1.9",
-    "react-native-screens": "~2.15.2",
     "react-native-unimodules": "^0.13.0",
     "react-native-vector-icons": "^8.1.0",
     "react-native-web": "^0.15.0",
@@ -52,6 +45,10 @@
     "metro-react-native-babel-preset": "^0.64.0",
     "react-test-renderer": "17.0.1",
     "typescript": "^3.8.3"
+  },
+  "resolutions": {
+    "node-fetch": "2.6.1",
+    "xmldom": "0.5.0"
   },
   "jest": {
     "preset": "react-native"
