+++ src/Components/Dashboard/Dashboard.tsx
@@ -1,19 +1,11 @@
 import React, { Component } from "react";
-import { Image, StyleSheet, Text, View } from "react-native";
-
-interface Props {
-  cards: any;
-}
-// interface State {
-//   uid: "",
-// }
-export default class Dashboard extends Component<Props> {
-  // constructor() {
-  //   super();
+import { Image, Text, View } from "react-native";
+import { DashboardProps, DashboardState } from "../../Utils/PropsState";
+import { styles } from "./DashboardStyles";
+export default class Dashboard extends Component<DashboardProps, DashboardState> {
   state = {
     uid: "",
   };
-  // }
 
   like = () => {};
 
@@ -45,45 +37,3 @@ export default class Dashboard extends Component<Props> {
     );
   }
 }
-
-const styles = StyleSheet.create({
-  container: {
-    width: "100%",
-    height: "100%",
-    padding: 20,
-    display: "flex",
-    justifyContent: "center",
-    alignItems: "center",
-  },
-  card: {
-    backgroundColor: "#fff",
-    borderRadius: 10,
-    paddingLeft: 20,
-    paddingRight: 20,
-    paddingTop: 20,
-    paddingBottom: 20,
-    justifyContent: "center",
-    alignItems: "center",
-
-    shadowColor: "#000",
-    shadowOpacity: 0.8,
-    shadowRadius: 2,
-    shadowOffset: {
-      height: 1,
-      width: 1,
-    },
-  },
-  title: {
-    fontSize: 24,
-  },
-  imageStyle: {
-    height: "60vh",
-    width: "75vw",
-    resizeMethod: "resize",
-  },
-  icon: {
-    justifyContent: "center",
-    alignItems: "center",
-    borderRadius: 100,
-  },
-});
