+++ src/Components/Dashboard/DashboardStyles.ts
@@ -1,0 +1,43 @@
+import { StyleSheet } from "react-native";
+
+export const styles = StyleSheet.create({
+  container: {
+    width: "100%",
+    height: "100%",
+    padding: 20,
+    display: "flex",
+    justifyContent: "center",
+    alignItems: "center",
+  },
+  card: {
+    backgroundColor: "#fff",
+    borderRadius: 10,
+    paddingLeft: 20,
+    paddingRight: 20,
+    paddingTop: 20,
+    paddingBottom: 20,
+    justifyContent: "center",
+    alignItems: "center",
+
+    shadowColor: "#000",
+    shadowOpacity: 0.8,
+    shadowRadius: 2,
+    shadowOffset: {
+      height: 1,
+      width: 1,
+    },
+  },
+  title: {
+    fontSize: 24,
+  },
+  imageStyle: {
+    height: "60vh",
+    width: "75vw",
+    resizeMethod: "resize",
+  },
+  icon: {
+    justifyContent: "center",
+    alignItems: "center",
+    borderRadius: 100,
+  },
+});
