+++ src/Components/Error/ErrorStyles.ts
@@ -1,0 +1,3 @@
+import { StyleSheet } from "react-native";
+
+export const styles = StyleSheet.create({});
